requirements.txt
Django==1.7.1
boto==2.34.0
django-allauth==0.18.0
django-debug-toolbar==1.2.2
djangorestframework==3.0.0
oauthlib==0.7.2
psycopg2==2.5.4
python-openid==2.2.5
requests==2.5.0
requests-oauthlib==0.4.2
sqlparse==0.1.14
wsgiref==0.1.2
